[MODIFY] 회원가입 UI 출력순서 수정
