+++ src/pages/User/Register.jsx
@@ -3,11 +3,10 @@ import Step1 from "../../components/User/Step1";
 import Step2 from "../../components/User/Step2";
 import Step3 from "../../components/User/Step3";
 import Step4 from "../../components/User/Step4";
-import Step5 from "../../components/User/Step5";
+import Step5 from "../../components/User/Step4.jsx";
 import { RegisterProvider, useRegister } from "../../components/User/RegisterContext";
 // 디버깅용 컴포넌트
 import { useLocation } from "react-router-dom";
-import DebugInfo from "../../components/User/DebugInfo.jsx";
 
 const steps = [<Step1 />, <Step2 />, <Step3 />, <Step4 />, <Step5 />];
 
@@ -20,7 +19,7 @@ const StepRenderer = () => {
     return (
         <>
             {steps[step - 1] || null}
-            <DebugInfo debugInfo={debugInfo} />
+            {/*<DebugInfo debugInfo={debugInfo} />*/}
         </>
     );
 };
